requirements.txt
Django==2.2.4
django-safedelete==0.5.2
djangorestframework==3.10.2
Django==2.2.4
django-safedelete==0.5.2
djangorestframework==3.10.2
pytz==2019.3
sqlparse==0.3.0
Django==2.2.4
django-safedelete==0.5.2
djangorestframework==3.10.2
mysqlclient==1.4.6
pytz==2019.3
sqlparse==0.3.0
